replace warn and error loggers
