+++ pkg/awsutil/util.go
@@ -6,25 +6,9 @@ import (
 	"net/http/httputil"
 	"regexp"
 
+	"github.com/rebuy-de/aws-nuke/pkg/util"
 	log "github.com/sirupsen/logrus"
 )
-
-func Indent(s, prefix string) string {
-	return string(IndentBytes([]byte(s), []byte(prefix)))
-}
-
-func IndentBytes(b, prefix []byte) []byte {
-	var res []byte
-	bol := true
-	for _, c := range b {
-		if bol && c != '\n' {
-			res = append(res, prefix...)
-		}
-		res = append(res, c)
-		bol = c == '\n'
-	}
-	return res
-}
 
 var REAuthHeader = regexp.MustCompile(`(?m:^(Auth[^:]*):.*$)`)
 
@@ -38,7 +22,7 @@ func DumpRequest(r *http.Request) string {
 
 	dump = bytes.TrimSpace(dump)
 	dump = REAuthHeader.ReplaceAll(dump, []byte("$1: <hidden>"))
-	dump = IndentBytes(dump, []byte("    > "))
+	dump = util.IndentBytes(dump, []byte("    > "))
 	return string(dump)
 }
 
@@ -51,6 +35,6 @@ func DumpResponse(r *http.Response) string {
 	}
 
 	dump = bytes.TrimSpace(dump)
-	dump = IndentBytes(dump, []byte("    < "))
+	dump = util.IndentBytes(dump, []byte("    < "))
 	return string(dump)
 }
